[RNE Rewrite] refactor(ocr): second review round on the OCR surface
ops/quad.ts is now just the quad type and the operations on it:
- Quad no longer claims its corners are ordered. The return type was lying
  wherever a detector handed one back, so callers had to know out of band which
  quads had been through orderQuad. Helpers that need the order say so.
- boundsOfPoints -> boundingBoxOfPoints, mapQuadToImage -> scaleQuad with a
  from/to/resizeMode options object mirroring scaleBox.
- rectifyQuad moves in from ops/image. restrictToBox is the precedent: a native
  op that takes a geometry type lives with that type.
- quadsFromFlat, splitWideQuad and orderByReadingOrder move out of the public
  API. The first is now private to paddleOcrUtils, the other two to paddleOcr.

paddleOcr.ts: inline property docs on OcrDetection, PaddleOcrModelOptions and
PaddleOcrModel, and defaultConfidenceThreshold is required now that it is the
documented default rather than something the task substitutes.

models.ts: PPOCRV6_OPTS untyped and inline, registry comment dropped, and only
PaddleOcrModel imported. index.ts stops re-exporting Quad, which is still
reachable as cv.Quad.

C++: rectifyQuad's locks sit next to fromJs like every other op in the file,
and two comments lose their trailing half.
